+++ src/components.tsx
@@ -2,7 +2,7 @@
  * @author: tisfeng
  * @createTime: 2022-06-26 11:13
  * @lastEditor: tisfeng
- * @lastEditTime: 2022-07-04 18:19
+ * @lastEditTime: 2022-07-05 12:48
  * @fileName: components.tsx
  *
  * Copyright (c) 2022 by tisfeng, All Rights Reserved.
@@ -60,7 +60,6 @@ export default function ListActionPanel(props: ActionListPanelProps) {
   checkIsInstalledEudic(setIsInstalledEudic);
 
   currentEasydict.getCurrentVersionInfo().then((easydict) => {
-    console.log(`has prompted: ${easydict.hasPrompted}`);
     setHasPrompted(easydict.hasPrompted);
   });